+++ tests/ImboClient/Url/MetadataTest.php
@@ -40,7 +40,7 @@ namespace ImboClient\Url;
  */
 class MetadataTest extends \PHPUnit_Framework_TestCase {
     /**
-     * Data provider for testGetUrl()
+     * Fetch URL data
      *
      * @return array
      */
@@ -51,12 +51,14 @@ class MetadataTest extends \PHPUnit_Framework_TestCase {
     }
 
     /**
+     * The metadata URL must be able to generate a complete URL with an access token appended
+     *
      * @dataProvider getUrlData
      * @covers ImboClient\Url\Url::getUrl
      * @covers ImboClient\Url\Metadata::getResourceUrl
      * @covers ImboClient\Url\Metadata::__construct
      */
-    public function testGetUrl($host, $publicKey, $image, $expected) {
+    public function testCanGenerateACompleteUrlIncludingAnAccessToken($host, $publicKey, $image, $expected) {
         $url = new Metadata($host, $publicKey, 'privateKey', $image);
         $this->assertStringStartsWith($expected, $url->getUrl());
         $this->assertRegExp('/accessToken=[a-f0-9]{64}$/', $url->getUrl());